# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/followers": {
            "post": {
                "tags": [
                    "Follower"
                ],
                "summary": "Подписаться на пользователя",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "follower_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вы успешно подписались на Clement Legros"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/followers/{user}": {
            "get": {
                "tags": [
                    "Follower"
                ],
                "summary": "Получить список подписок пользователя",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "follower_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "follower_name": {
                                                        "type": "string",
                                                        "example": "Gracie Sauer Sr."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/followers/subscriptions/{user}": {
            "get": {
                "tags": [
                    "Follower"
                ],
                "summary": "Получить список подписчиков пользователя",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id пользователя",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "subscription_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "subscription_name": {
                                                        "type": "string",
                                                        "example": "Gracie Sauer Sr."
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/followers/{follower}": {
            "delete": {
                "tags": [
                    "Follower"
                ],
                "summary": "Отписка от пользователя",
                "parameters": [
                    {
                        "name": "follower",
                        "in": "path",
                        "description": "Id подписки",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вы успешно отписались от пользователя"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}